requirements.txt
streamlit
opencv-python-headless
rembg
numpy
Pillow
streamlit-cropper
pandas
watchdog
